port should be uint16_t, not long

tested by areynold

+++ commotion-service-manager.c
@@ -358,7 +358,7 @@ void resolve_callback(
     const char *domain,
     const char *host_name,
     const AvahiAddress *address,
-    long port,
+    uint16_t port,
     AvahiStringList *txt,
     AvahiLookupResultFlags flags,
     void* userdata) {
